src/main/java/ru/yandex/practicum/filmorate/storage/impl/dao/DbMpaStorage.java
package ru.yandex.practicum.filmorate.storage.impl.dao;

import lombok.RequiredArgsConstructor;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.stereotype.Repository;
import ru.yandex.practicum.filmorate.model.Mpa;
import ru.yandex.practicum.filmorate.storage.MpaStorage;

import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.List;

@Repository("ratingMpaDbStorage")
@RequiredArgsConstructor
public class DbMpaStorage implements MpaStorage {

    private final JdbcTemplate jdbcTemplate;

    @Override
    public List<Mpa> findAll() {
        return jdbcTemplate.query("SELECT * FROM rating;", this::makeRatingMpa);
    }

    @Override
    public Mpa findById(Integer id) {
        return jdbcTemplate.queryForObject("SELECT * FROM rating WHERE id = ?", this::makeRatingMpa, id);
    }

    private Mpa makeRatingMpa(ResultSet resultSet, int rowNum) throws SQLException {
        return new Mpa(resultSet.getInt("id"), resultSet.getString("name"));
    }
}
